remove lint errors
