+++ src/app/[locale]/(main-public)/_components/LandingPage/AdmissionOpen.tsx
@@ -1,12 +1,13 @@
+import { getScopedI18n } from "@/locales/server";
 import { Button } from "@lib/next-ui";
-import type { FC } from "react";
 import Image from "next/image";
 import Link from "next/link";
-import { getScopedI18n } from "@/locales/server";
+import type { FC } from "react";
 
-const AdmissionOpen: FC =async () => {
+const AdmissionOpen: FC = async () => {
   const currYear = new Date().getFullYear();
   const t = await getScopedI18n("Root.LandingPage.admission");
+
   return (
     <section className="grid grid-cols-1 gap-6 rounded-xl bg-primary-400 p-8 text-slate-50 sm:grid-flow-row sm:grid-cols-2 md:p-10 lg:grid-flow-col lg:grid-cols-3 lg:items-center lg:gap-x-16">
       <h3 className="order-1 text-3xl font-bold lg:self-end lg:text-4xl">
