--- src/app/[locale]/(main-public)/_components/LandingPage/Index.tsx
@@ -1,40 +1,0 @@
-
-import React, {type FC } from "react";
-import MainCarosal from "./HeroCarousel";
-import AboutUs from "./AboutUs";
-import PhotoGallery from "./PhotoGallery";
-import KeyCharacteristics from "./KeyCharacteristics";
-import AdmissionOpen from "./AdmissionOpen";
-import PremiumFacilities from "./PremiumFacilities";
-
-const Index:FC = () => {
-  return (
-    <>
-      <section>
-        <MainCarosal />
-      </section>
-
-      <section>
-        <AboutUs />
-      </section>
-
-      <section>
-        <PhotoGallery />
-      </section>
-
-      <section>
-        <KeyCharacteristics />
-      </section>
-
-      <section>
-        <AdmissionOpen />
-      </section>
-
-      <section>
-        <PremiumFacilities />
-      </section>
-    </>
-  );
-};
-
-export default Index;
