+++ src/app/[locale]/(main-public)/_components/Navbar/SideNavbar.tsx
@@ -1,10 +1,10 @@
-import type { NAV_LINK_TYPES } from "@/lib/frontend-data";
-import { NAV_LINKS } from "@/lib/frontend-data";
 import { Button, Divider } from "@lib/next-ui";
 
+import type { NAV_LINK_TYPES } from "@/lib/frontend-data";
+import { NAV_LINKS } from "@/lib/frontend-data";
 import { getScopedI18n } from "@/locales/server";
 import NextLink from "next/link";
-import { type FC } from "react";
+import type { FC } from "react";
 
 export type Props = {
   linkType: (typeof NAV_LINK_TYPES)[number];
