--- src/locales/en.ts
@@ -1,309 +1,0 @@
-export default {
-  Dictionary: {
-    other: "other",
-    links: "links",
-  },
-  Pages: {
-    school: {
-      title: "School",
-      "sub-links": {
-        about: {
-          title: "About Us",
-          content: {
-            "para-1": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Nunc faucibus a pellentesque sit amet porttitor eget dolor morbi. Lorem donec massa sapien faucibus. Leo vel fringilla est ullamcorper eget nulla. Accumsan tortor posuere ac ut. Rutrum quisque non tellus orci ac auctor augue mauris augue. Nunc faucibus a pellentesque sit. Elementum tempus egestas sed sed risus pretium. Sit amet luctus venenatis lectus. Ut pharetra sit amet aliquam id diam. Eu mi bibendum neque egestas congue quisque egestas. Quisque non tellus orci ac auctor augue mauris. Erat nam at lectus urna duis convallis. Quis hendrerit dolor magna eget est lorem ipsum dolor sit. Quis vel eros donec ac odio tempor orci. Ultrices dui sapien eget mi proin sed libero enim sed.",
-            "para-2": "Ut sem viverra aliquet eget sit amet tellus. Et egestas quis ipsum suspendisse ultrices gravida. Aenean sed adipiscing diam donec adipiscing tristique risus nec. Morbi tristique senectus et netus et. Lectus vestibulum mattis ullamcorper velit sed ullamcorper morbi. Convallis tellus id interdum velit laoreet id donec ultrices tincidunt. Mollis aliquam ut porttitor leo a diam sollicitudin tempor id. Nam at lectus urna duis convallis. Vulputate mi sit amet mauris commodo quis imperdiet massa tincidunt. Non pulvinar neque laoreet suspendisse interdum.",
-          },
-        },
-        mission: {
-          title: "Mission and Vision",
-          content: {
-            "para-1": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Nunc faucibus a pellentesque sit amet porttitor eget dolor morbi. Lorem donec massa sapien faucibus. Leo vel fringilla est ullamcorper eget nulla. Accumsan tortor posuere ac ut. Rutrum quisque non tellus orci ac auctor augue mauris augue. Nunc faucibus a pellentesque sit. Elementum tempus egestas sed sed risus pretium. Sit amet luctus venenatis lectus. Ut pharetra sit amet aliquam id diam. Eu mi bibendum neque egestas congue quisque egestas. Quisque non tellus orci ac auctor augue mauris. Erat nam at lectus urna duis convallis. Quis hendrerit dolor magna eget est lorem ipsum dolor sit. Quis vel eros donec ac odio tempor orci. Ultrices dui sapien eget mi proin sed libero enim sed.",
-            "para-2": "Ut sem viverra aliquet eget sit amet tellus. Et egestas quis ipsum suspendisse ultrices gravida. Aenean sed adipiscing diam donec adipiscing tristique risus nec. Morbi tristique senectus et netus et. Lectus vestibulum mattis ullamcorper velit sed ullamcorper morbi. Convallis tellus id interdum velit laoreet id donec ultrices tincidunt. Mollis aliquam ut porttitor leo a diam sollicitudin tempor id. Nam at lectus urna duis convallis. Vulputate mi sit amet mauris commodo quis imperdiet massa tincidunt. Non pulvinar neque laoreet suspendisse interdum.",
-          },
-        },
-        principal: {
-          title: "Principal's Message",
-          content: {
-            avatar: "Prinicpal of Greenfield",
-            "para-1": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Nunc faucibus a pellentesque sit amet porttitor eget dolor morbi. Lorem donec massa sapien faucibus. Leo vel fringilla est ullamcorper eget nulla. Accumsan tortor posuere ac ut. Rutrum quisque non tellus orci ac auctor augue mauris augue. Nunc faucibus a pellentesque sit. Elementum tempus egestas sed sed risus pretium. Sit amet luctus venenatis lectus. Ut pharetra sit amet aliquam id diam. Eu mi bibendum neque egestas congue quisque egestas. Quisque non tellus orci ac auctor augue mauris. Erat nam at lectus urna duis convallis. Quis hendrerit dolor magna eget est lorem ipsum dolor sit. Quis vel eros donec ac odio tempor orci. Ultrices dui sapien eget mi proin sed libero enim sed.",
-            "para-2": "Ut sem viverra aliquet eget sit amet tellus. Et egestas quis ipsum suspendisse ultrices gravida. Aenean sed adipiscing diam donec adipiscing tristique risus nec. Morbi tristique senectus et netus et. Lectus vestibulum mattis ullamcorper velit sed ullamcorper morbi. Convallis tellus id interdum velit laoreet id donec ultrices tincidunt. Mollis aliquam ut porttitor leo a diam sollicitudin tempor id. Nam at lectus urna duis convallis. Vulputate mi sit amet mauris commodo quis imperdiet massa tincidunt. Non pulvinar neque laoreet suspendisse interdum.",
-          },
-        },
-        infrastructure: {
-          title: "Infrastructure",
-          content: {},
-        },
-        achievements: {
-          title: "Achievements",
-          content: {},
-        },
-        gallery: {
-          title: "Gallery",
-          content: {},
-        },
-        news: {
-          title: "In the News",
-          content: {},
-        },
-        transport: {
-          title: "School Transport",
-          content: {},
-        },
-        pledge: {
-          title: "Our Pledge",
-          content: {
-            avatar: "Prinicpal of Greenfield",
-            "para-1": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Nunc faucibus a pellentesque sit amet porttitor eget dolor morbi. Lorem donec massa sapien faucibus. Leo vel fringilla est ullamcorper eget nulla. Accumsan tortor posuere ac ut. Rutrum quisque non tellus orci ac auctor augue mauris augue. Nunc faucibus a pellentesque sit. Elementum tempus egestas sed sed risus pretium. Sit amet luctus venenatis lectus. Ut pharetra sit amet aliquam id diam. Eu mi bibendum neque egestas congue quisque egestas. Quisque non tellus orci ac auctor augue mauris. Erat nam at lectus urna duis convallis. Quis hendrerit dolor magna eget est lorem ipsum dolor sit. Quis vel eros donec ac odio tempor orci. Ultrices dui sapien eget mi proin sed libero enim sed.",
-            "para-2": "Ut sem viverra aliquet eget sit amet tellus. Et egestas quis ipsum suspendisse ultrices gravida. Aenean sed adipiscing diam donec adipiscing tristique risus nec. Morbi tristique senectus et netus et. Lectus vestibulum mattis ullamcorper velit sed ullamcorper morbi. Convallis tellus id interdum velit laoreet id donec ultrices tincidunt. Mollis aliquam ut porttitor leo a diam sollicitudin tempor id. Nam at lectus urna duis convallis. Vulputate mi sit amet mauris commodo quis imperdiet massa tincidunt. Non pulvinar neque laoreet suspendisse interdum.",
-          },
-        },
-        "news-letter": {
-          title: "Newsletter",
-          content: {},
-        },
-        admissions: {
-          title: "Admissions",
-          content: {},
-        },
-        contact: {
-          title: "Contact Us",
-          content: {},
-        },
-      },
-    },
-    admission: {
-      title: "Admissions",
-      "sub-links": {
-        notice: {
-          title: "Admission Notice",
-          content: {},
-        },
-        prospectus: {
-          title: "Prospectus",
-          content: {},
-        },
-        procedure: {
-          title: "Admission Procedure",
-          content: {},
-        },
-        instructions: {
-          title: "General Instructions for Admission",
-          content: {},
-        },
-        age: {
-          title: "Age Criteria",
-          content: {},
-        },
-        fee: {
-          title: "Fee Structure",
-          content: {},
-        },
-        orientation: {
-          title: "Orientation",
-          content: {},
-        },
-        registration: {
-          title: "Online Registration",
-          content: {},
-        },
-      },
-    },
-    academic: {
-      title: "Academics",
-      "sub-links": {
-        result: {
-          title: "Result Analysis",
-          content: {},
-        },
-        syllabus: {
-          title: "Syllabus",
-          content: {},
-        },
-        "book-list": {
-          title: "Book List",
-          content: {},
-        },
-        holiday: {
-          title: "Holiday List",
-          content: {},
-        },
-        "block-holidays": {
-          title: "Block Holidays",
-          content: {},
-        },
-        fees: {
-          title: "Fee Structure",
-          content: {},
-        },
-      },
-    },
-    student: {
-      title: "Students",
-      "sub-links": {
-        portal: {
-          title: "Student Portal",
-          content: {},
-        },
-        council: {
-          title: "Student Council",
-          content: {},
-        },
-        activities: {
-          title: "Co-Curricular Activities",
-          content: {},
-        },
-        timings: {
-          title: "School Timings",
-          content: {},
-        },
-        certificate: {
-          title: "Transfer Certificate",
-          content: {},
-        },
-        prayer: {
-          title: "School Prayer",
-          content: {},
-        },
-        trips: {
-          title: "Trips & Excursions",
-          content: {},
-        },
-        cc: {
-          title: "Code of Conduct",
-          content: {},
-        },
-        "s-rules": {
-          title: "Special Rules",
-          content: {},
-        },
-        "l-rules": {
-          title: "Library Rules",
-          content: {},
-        },
-        "c-rules": {
-          title: "Canteen Rules",
-          content: {},
-        },
-        "+ve-behavior-manage": {
-          title: "Positive Behavior Management",
-          content: {},
-        },
-        cce: {
-          title: "Continuous and Comprehensive Evaluation",
-          content: {},
-        },
-      },
-    },
-    parent: {
-      title: "Parents",
-      "sub-links": {
-        dispersal: {
-          title: "Student's Dispersal",
-          content: {},
-        },
-        uniform: {
-          title: "School Uniform",
-          content: {},
-        },
-        hours: {
-          title: "Visiting Hours",
-          content: {},
-        },
-        guidelines: {
-          title: "General Guidelines",
-          content: {},
-        },
-        absence: {
-          title: "Absence - Leave - Late Arrival in School",
-          content: {},
-        },
-        infirmary: {
-          title: "School Infirmary",
-          content: {},
-        },
-        report: {
-          title: "Progress Report",
-          content: {},
-        },
-        fee: {
-          title: "Pay Fee Online",
-          content: {},
-        },
-        login: {
-          title: "Parent's Login",
-          content: {},
-        },
-      },
-    },
-  },
-  Root: {
-    LocaleSwitcher: {
-      label: "Change language",
-      locale: "{locale, select, ta {Tamil} te {Telugu} en {English} hi {Hindi} other {Unknown}}",
-    },
-    "main-navbar": {
-      "admission-portal": "Admission Portal",
-      login: "Login",
-    },
-    LandingPage: {
-      hero: {
-        title: "Welcome to Greenfield International School",
-        "para-1": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Nunc faucibus a pellentesque sit amet porttitor eget dolor morbi. Lorem donec massa sapien faucibus. Leo vel fringilla est ullamcorper eget nulla. Accumsan tortor posuere ac ut. Rutrum quisque non tellus orci ac auctor augue mauris augue. Nunc faucibus a pellentesque sit. Elementum tempus egestas sed sed risus pretium. Sit amet luctus venenatis lectus. Ut pharetra sit amet aliquam id diam. Eu mi bibendum neque egestas congue quisque egestas. Quisque non tellus orci ac auctor augue mauris. Erat nam at lectus urna duis convallis. Quis hendrerit dolor magna eget est lorem ipsum dolor sit. Quis vel eros donec ac odio tempor orci. Ultrices dui sapien eget mi proin sed libero enim sed.",
-        "about-us": "About us",
-      },
-      photogallery: {
-        title: "Photo Gallery",
-        cardElements: {
-          first: {
-            heading: "What to watch",
-            subHeading: "Stream the Acme event",
-          },
-          second: {
-            heading: "Plant a tree",
-            subHeading: "Contribute to the planet",
-          },
-          third: {
-            heading: "Supercharged",
-            subHeading: "Creates beauty like a beast",
-          },
-          forth: {
-            heading: "Super-duper-charged",
-            subHeading: "Creates beauty like a beast",
-          },
-        },
-      },
-
-      admission: {
-        title: "School Admission Open",
-        button: "Enroll Now",
-        para1: "Get your kid the best experience of education at Greenfield School,Tamil Nadu. Best for your child ",
-      },
-
-      characteristics: {
-        title: " Key differentiating characteristics of the School",
-        cardElements: {
-          first: {
-            heading: "Tomorrow’s Leaders",
-            body: "Here with us, students develop a sense of self, social skills, and independence, preparing them to take on leadership roles in the future."
-          },
-          second: {
-            heading: "Noteworthy technology acquisitions 2021",
-            body: "Here are the biggest enterprise technology acquisitions of 2021 so far, in reverse chronological order.",
-          },
-          third: {
-            heading: "Tomorrow’s Leaders are",
-            body: "Here with us, students develop a sense of self, social skills, and independence, preparing them to take on leadership roles in the future.",
-          },
-          forth: {
-            heading: "Tomorrow’s Leaders good",
-            body: "Here with us, students develop a sense of self, social skills, and independence, preparing them to take on leadership roles in the future.",
-          },
-        },
-      },
-
-      premiunFacilities :{
-        title: "Premium Facilities"
-      }
-    },
-  },
-} as const;
