+++ src/locales/languages/hi.ts
