+++ src/locales/languages/ta.ts
