+++ src/locales/languages/te.ts
