+++ src/locales/server.ts
@@ -2,8 +2,8 @@ import { createI18nServer } from "next-international/server";
 
 export const { getI18n, getScopedI18n, getStaticParams, getCurrentLocale } =
   createI18nServer({
-    en: () => import("./en"),
-    hi: () => import("./hi"),
-    ta: () => import("./ta"),
-    te: () => import("./te"),
+    en: async () => await import("./languages/en"),
+    hi: async () => await import("./languages/hi"),
+    ta: async () => await import("./languages/ta"),
+    te: async () => await import("./languages/te"),
   });
