--- src/types/global.d.ts
@@ -1,5 +1,0 @@
-// Use type safe message keys with `next-intl`
-import type Message from "../messages/en.json";
-
-type Messages = typeof Message;
-declare interface IntlMessages extends Messages {}
